input-service/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
py-solc-x
python-dotenv
requests